Add UUID support
