+++ phpunit.php
@@ -10,3 +10,9 @@ $capsule->setAsGlobal();
 
 include __DIR__.'/tests/models/Category.php';
 include __DIR__.'/tests/models/MenuItem.php';
+include __DIR__.'/tests/ScopedNodeTestBase.php';
+include __DIR__.'/tests/NodeTestBase.php';
+include __DIR__.'/tests/models/CategoryUuid.php';
+include __DIR__.'/tests/models/MenuItemUuid.php';
+include __DIR__.'/tests/data/CategoryData.php';
+include __DIR__.'/tests/data/MenuItemData.php';
