+++ src/NestedSet.php
@@ -36,11 +36,16 @@ class NestedSet
      *
      * @param \Illuminate\Database\Schema\Blueprint $table
      */
-    public static function columns(Blueprint $table)
+    public static function columns(Blueprint $table, bool $uuid = false)
     {
         $table->unsignedInteger(self::LFT)->default(0);
         $table->unsignedInteger(self::RGT)->default(0);
-        $table->unsignedInteger(self::PARENT_ID)->nullable();
+
+        if ($uuid) {
+            $table->uuid(self::PARENT_ID)->nullable()->index();
+        } else {
+            $table->unsignedInteger(self::PARENT_ID)->nullable();
+        }
 
         $table->index(static::getDefaultColumns());
     }
