+++ tests/NodeUuidTest.php
@@ -2,29 +2,29 @@
 
 use Kalnoy\Nestedset\NestedSet;
 
-class ScopedNodeTest extends ScopedNodeTestBase
+class NodeUuidTest extends NodeTestBase
 {
     public function __construct($name = null)
     {
         parent::__construct($name);
-        $this->menuItemData = new MenuItemData();
+        $this->categoryData = new CategoryData();
     }
 
     protected function getTable(): string
     {
-        return 'menu_items';
+        return 'uuid_categories';
     }
 
     protected function getModelClass(): string
     {
-        return MenuItem::class;
+        return CategoryUuid::class;
     }
 
     protected function createTable(\Illuminate\Database\Schema\Blueprint $table): void
     {
-        $table->increments('id');
-        $table->unsignedInteger('menu_id');
-        $table->string('title')->nullable();
-        NestedSet::columns($table);
+        $table->uuid('id')->primary();
+        $table->string('name');
+        $table->softDeletes();
+        NestedSet::columns($table, true);
     }
 }
