+++ tests/models/MenuItemUuid.php
@@ -1,0 +1,12 @@
+<?php
+
+class MenuItemUuid extends MenuItem
+{
+    use \Illuminate\Database\Eloquent\Concerns\HasUuids;
+
+    protected $table = 'uuid_menu_items';
+
+    protected $keyType = 'string';
+
+    public $incrementing = false;
+}
